check for gitconfig before mv

+++ install.sh
@@ -77,7 +77,7 @@ if [ $SPIN ]; then
   # Set system generated .gitconfig to .gitconfig.local. We'll pull it in later as part
   # of our custom gitconfig. The generated gitconfig already has the right user and email,
   # since Spin knows that from login.
-  mv -n $HOME/.gitconfig $HOME/.gitconfig.local
+  [ -f $HOME/.gitconfig ] && mv -n $HOME/.gitconfig $HOME/.gitconfig.local
 
   echo '👌 finito, spin setup'
 fi
